Hide sales form when submitted
Previously, when the contact form was submitted, the page scrolled to
the top and the success notification was plain text that looked the same
as other text on the page. This meant potential customers don’t know if
the form has successfully submitted, so they often resubmit multiple
times. At best this produces a lot of noise in the SDR slack room and
bad metrics. At worst two different SDRs can contact the same lead 5
minutes after each other.

This commit changes the UX. When the form successfully submits it
disappears and is replaced by a bright notice box that says
’Thanks. We’ll be in touch soon.’ or something to that effect.