Remove empty href attributes
Turning `<a href=''>..` into `<a>..`

Having an empty href seems to navigate to root, triggering a full page reload. This is bad because we used empty href where we wanted js to trigger a dropdown to open, it was impossible to open a dropdown and then click on something in it.

This is currently not a problen in production but is happening on circle and locally. Not sure why.