Update example checkout to v2
Current links to checkout demo pages open the v1 payment pages. This change updates the link order (subscriptions first, as this is the most relevant one) and opens the new payment pages in a new window.

Note: The old payment links opened pre-filled forms (ie. with account number & sort code). This is not possible with the new payment pages anymore, we should reconsider how we can improve the whole demo concept.